ames, gall: halt ames flow if %flub
a previous version would inmediately cork the flow if the flubbed $plea was a leave. this introduces complexity that is not worth the possibility of entering into a de-sync state. in the current approach we just halt the flow, dropping any incoming messages, and queueing and outstanding, without sending them