hoon: consistently support empty $pith segments
Path literal syntax supports writing either $ or nothing at all for an
empty segment (though $ is needed for a _trailing_ empty segment). Pith
literal syntax supported $, but not the omission of a segment body
entirely. Additionally, pith pattern syntax did not support either.

Here, we allow parsing the empty string as a %$ path segment in all
three parsers, and fix parsing of the $ segment in pattern syntax.