ames: change $path in %ahoy $plea
in anticipation og the %mass-ahoy that will migrate everybody to use the directed messaging protocol, implemented in the |mesa core in the  %ames vane.

to account for peers that are behind receiving the OTA that would enable %mass-mate (which will include an update migration logic) we update the path in the %ahoy $plea to trigger a crash on the receiver if it doesn't handle the new code.

(NOTE: for mass migrations we need to add a retry functionality. this can be added to user space or to %ames by tracking nacked %ahoy flows)