ames, gall: don't break exactly-once message delivery
(note: WIP the in-progress %pleas are never cleared/acked and they keep being resent)

if a %plea has been saved in the blocked queue in %gall, we don't want to %flub in %ames since that will re-deliver the same message again to the agent.

The first time the message is delivered is when the agent is installed/revived, then when the resend of the last packet arrives in ames (we start from the "in-progress state"), and if we have done the "normal" flub, we would make it look like the message was never delivered by decreasing last-heard and removing the pending-vane-ack.